requirements.txt
Appium-Python-Client==1.3.0
selenium==3.141.0
urllib3<2.0
pytest==7.4.3
pyautogui==0.9.54
pyperclip==1.8.2
Flask==3.0.0
uiautomation
requests>=2.31.0
Pillow>=10.4.0
pywin32>=306

